// src/interface/theme.type.ts
// src/interface/theme.type.ts

export interface themeType {
  // Base Background
  backgroundColor: string;
  backgroundColorS1: string;
  backgroundColorS2: string;

  // Brand Colors
  primary: string;
  primaryS1: string;
  primaryS2: string;
  primaryS3: string;
  secondary: string;
  tertiary: string;

  // Text
  text: string;
  textS1: string;
  textS2: string;
  descriptionText: string;
  extraLightText: string;
  lightTextBlack: string;
  lightTextwhite: string;

  // Borders / Placeholder
  borderColor: string;
  placeHolderColor: string;

  // Utilities
  white: string;
  black: string;
  red: string;
  redS1: string;
  gray: string;
  grayS1: string;
  grayS2: string;
  grayS3: string;
  green: string;
  greenS1: string;
  brown: string;
  yellow: string;
  yellowS1: string;
  lightOrange: string;
}